+++ src/main/kotlin/com/wutsi/codegen/kotlin/server/ServerCodeGenerator.kt
@@ -3,6 +3,7 @@ package com.wutsi.codegen.kotlin.server
 import com.wutsi.codegen.Context
 import com.wutsi.codegen.core.generator.CodeGenerator
 import com.wutsi.codegen.editconfig.EditorConfigCodeGenerator
+import com.wutsi.codegen.github.GithubWorkflowCodeGenerator
 import com.wutsi.codegen.kotlin.KotlinMapper
 import com.wutsi.codegen.kotlin.sdk.SdkModelCodeGenerator
 import io.swagger.v3.oas.models.OpenAPI
@@ -15,5 +16,6 @@ class ServerCodeGenerator(private val mapper: KotlinMapper) : CodeGenerator {
         ServerPomCodeGenerator(mapper).generate(openAPI, context)
         ServerLauncherCodeGenerator().generate(openAPI, context)
         EditorConfigCodeGenerator().generate(openAPI, context)
+        GithubWorkflowCodeGenerator().generate(openAPI, context)
     }
 }